Create middleware folder to store all middleware files. Extract out authenticate_token middleware and create a file for it. Deploy this middleware to model.js
